refactor: Room 관련 클래스 패키지 경로 이동 (local -> local.room)
